Include missing function on header file
There is an issue when installing this gem on recent MacOs:

Clang now reports an error when you use a function without an explicit declaration when building C or Objective-C code for macOS (-Werror=implicit-function-declaration flag is on). This additional error detection unifies Clang’s behavior for iOS/tvOS and macOS 64-bit targets for this diagnostic. (49917738)

https://github.com/tmm1/rbtrace/issues/81